fetch informations of created resources from the response message text
After an object is created via POST, the ID of the created object is only present in the reponse message text. Therefore a mechanism was implemented to extract the ID and type of a created resource.

Example for the message text of a created resource:
Resource [REALESTATE] with id [123456] has been created.