always export "xsd:date" values without timezone
Because of a bug in the IS24-Webservice some date values are ignored by the Webservice, if a timezone is present in the date string - according to the specifications of xsd:date (see http://www.w3schools.com/schema/schema_dtypes_date.asp). As a workaround to that issue, OpenEstate-IS24-REST does not append timezones to any exported xsd:date values anymore.

JAXB classes were regenerated in order to fix that issue.