process L-IS24-RequestRefnum header for failing requests
Store the value of "L-IS24-RequestRefnum" header for failing requests into the "RequestFailedException" (for low level exports) and "ExportMessage" (for bulk exports). In case of problems with a certain request, one can provide the corresponding number to IS24 together with a support request. This allows IS24 to track the error / problem much easier and to provide better support for their users. This resolves #2.

Besides this some API improvements & documentations were made.