fixed URL encoding
URL fragments are encoded with "java.net.URLEncoder". This class does not escape spaces as expected ("+" instead of "%20"). This may lead to problems, when a property is identified by its external ID and the external ID contains spaces. To fix these problems we should replace any "+" with "%20" in the escaped string.

see also https://www.subbu.org/blog/2008/02/uri-escaping-and-javaneturlencoder